Fix handling of MWM decoration hints (issue #100).
JWM was treating the decoration hints the same way as the function hints.
This made it impossible to minimize/maximize a window that requested that
the minimize/maximize buttons to not be shown.
Now JWM ignores the MWM decoration hints for these buttons, though JWM
still obeys the title and outline hints.