Add backandforth attribute to Desktops tag
This commit adds a backandforth attribute to the Desktops tag.
The accepted values are "on and "off" (the default).

When this attribute is set to "on" trying to change desktops to the
currently active desktop with the desktop# action will change to the
previously active desktop instead.

When it is set to "off" (or not set) the previous behavior of doing
nothing is retained.