Fix windowType not being initialized.
ReadWindowState() was creating ClientState structs without
initializing every field, so windowType was typically zero (might
not be: this is undefined behaviour I think). Zero is the same as
WINDOW_TYPE_DESKTOP, which was breaking my checks for whether a
window is a desktop window. For some reason, all normal windows
were initialized that way too.

Default it to WINDOW_TYPE_NORMAL, much like the defaultLayer is
defaulted to LAYER_NORMAL.