avoid continuous growth of memory usage in case of scaled images and xrender
In case of xrender and usage of scaled images JWM had growing memory
consumption due to allocation of new ScaledIconNode if icon is requested. For
instance, on each focus change the icon is requested and new ScaledIconNode had
been allocated. xrestop shows it as well as increasing pixmap bytes and amount
of pictures.

The reason is that CreateScaledRenderIcon returns unscaled icon and
PutScaledRenderIcon makes actual scaling. The lookup in list of scaled icons
assumed that all icons are already scaled and found nothing, then created new
one, put it into list and so on.

This commit corrects lookup of ScaledIconNode by adding additional fields to
handle post-scaling by xrender.